requirements.txt
more-itertools==10.1.0
PyYAML==6.0.1
